+++ hypo/src/test/java/com/z0ltan/compilers/vms/hypo/AppTest.java
@@ -18,18 +18,18 @@ public class AppTest extends TestCase {
   public void testAddnums() {
     final Interpreter interpreter = new Interpreter("samples/addnums.S");
     interpreter.emulate();
-    assertEquals(interpreter.result(), 5);
+    assertEquals(interpreter.result(), (short)5);
   }
 
   public void testSubnums() {
     final Interpreter interpreter = new Interpreter("samples/subnums.S");
     interpreter.emulate();
-    assertEquals(interpreter.result(), -1);
+    assertEquals(interpreter.result(), (short)-1);
   }
 
   public void testProduct() {
     final Interpreter interpreter = new Interpreter("samples/product.S");
     interpreter.emulate();
-    assertEquals(interpreter.result(), 6);
+    assertEquals(interpreter.result(), (short)6);
   }
 }
